requirements.txt
# ── LangChain / LangGraph ─────────────────────────────────────────────────────
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0

# ── Community extensions (HuggingFace embeddings for RAG) ────────────────────
langchain-community>=0.3.0

# ── Redis (SSE progress streaming + report caching) ───────────────────────────
redis>=4.6.0

# ── PDF generation ────────────────────────────────────────────────────────────
weasyprint>=61.0

# ── LLM / AI ─────────────────────────────────────────────────────────────────
pydantic>=2.0.0

# ── Embeddings (pinned for torch 2.2.x / numpy 1.x compatibility) ────────────
numpy<2.0
sentence-transformers<3.0.0

# ── pgvector RAG ──────────────────────────────────────────────────────────────
pgvector>=0.3.0
psycopg2-binary>=2.9.9

# ── HTTP & scraping ───────────────────────────────────────────────────────────
requests>=2.31.0
httpx>=0.27.0
beautifulsoup4>=4.12.0

# ── Environment ───────────────────────────────────────────────────────────────
python-dotenv>=1.0.0

# ── API server ────────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sse-starlette>=1.8.2
aiofiles>=23.0.0
